Bugfix user comment

+++ Packages/MIES/MIES_MiesUtilities.ipf
@@ -1656,6 +1656,12 @@ Function/WAVE GetSweepsWithSetting(labnotebookValues, setting)
 		Make/FREE/N=(DimSize(indizes, ROWS)) sweeps = labnotebookValues[indizes[p]][sweepCol][0]
 	endif
 
+	WAVE/Z cleanSweeps = ZapNaNs(sweeps)
+
+	if(!WaveExists(cleanSweeps))
+		return $""
+	endif
+
 	// @todo IP9: Make it threadsafe once FindDuplicates is threadsafe
 	return DeleteDuplicates(sweeps)
 End
@@ -1842,7 +1848,30 @@ Function/S GetAllDevices()
 	return list
 End
 
-/// @brief Returns a list of all devices, e.g. "ITC18USB_Dev_0;", which have acquired data.
+static Function DeviceHasUserComments(string device)
+
+	string userCommentDraft, userCommentNB, userComment, commentNotebook
+
+	userComment = ROStr(GetUserComment(device))
+
+	if(WindowExists(device))
+		userCommentDraft = DAG_GetTextualValue(device, "SetVar_DataAcq_Comment")
+
+		commentNotebook = DAP_GetCommentNotebook(device)
+		if(WindowExists(commentNotebook))
+			userCommentNB = GetNotebookText(commentNotebook)
+		else
+			userCommentNB = ""
+		endif
+	else
+		userCommentNB    = ""
+		userCommentDraft = ""
+	endif
+
+	return !IsEmpty(userComment) || !IsEmpty(userCommentDraft) || !IsEmpty(userCommentNB)
+End
+
+/// @brief Returns a list of all devices, e.g. "ITC18USB_Dev_0;", which have content.
 ///
 /// @param contentType [optional, defaults to CONTENT_TYPE_SWEEP] type of
 ///                    content to look for, one of @ref CONTENT_TYPES
@@ -1875,6 +1904,12 @@ Function/S GetAllDevicesWithContent([contentType])
 		if(contentType & CONTENT_TYPE_TPSTORAGE                                     \
 		   && DataFolderExists(testPulsePath)                                       \
 		   && ItemsInList(GetListOfObjects($testPulsePath, TP_STORAGE_REGEXP)) > 0)
+			list = AddListItem(device, list, ";", inf)
+			continue
+		endif
+
+		if(contentType & CONTENT_TYPE_COMMENT \
+		   && DeviceHasUserComments(device))
 			list = AddListItem(device, list, ";", inf)
 			continue
 		endif
